Added MessageID to CreateGameState publish
Added checks for GUIDs

+++ DealerBear/UseCases/GenerateSeed/Interface/IGenerateSeed.cs
@@ -2,6 +2,6 @@ namespace DealerBear.UseCases.GenerateSeed.Interface
 {
     public interface IGenerateSeed
     {
-        float Execute();
+        int Execute();
     }
 }
